# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: ashim API
  version: 1.15.9
  description: |
    REST API for ashim, a self-hosted image processing platform with 30+ tools.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Session cookie** — Call `POST /api/auth/login` with username and password. The response includes a `token` field. Pass it as `Authorization: Bearer <token>` on subsequent requests.

    2. **API key** — Generate a key via the Settings UI or `POST /api/v1/api-keys`. Keys are prefixed with `si_`. Pass as `Authorization: Bearer si_...`.

    Endpoints marked with a lock icon require authentication. Admin-only endpoints are noted in their description.
  license:
    name: AGPL-3.0
    url: https://github.com/ashim-hq/ashim/blob/main/LICENSE
servers:
  - url: /
    description: Current instance
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login or API key (prefixed with si_)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid image format
    ToolResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Unique job identifier
        downloadUrl:
          type: string
          description: URL to download the processed image
          example: /api/v1/download/abc123/output.png
        originalSize:
          type: integer
          description: Original file size in bytes
        processedSize:
          type: integer
          description: Processed file size in bytes
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Authentication required
...
    FeatureNotInstalledError:
      type: object
      properties:
        error:
          type: string
          example: Feature not installed
        code:
          type: string
          example: FEATURE_NOT_INSTALLED
        feature:
          type: string
          description: Feature bundle identifier
        featureName:
          type: string
          description: Human-readable feature name
        estimatedSize:
          type: string
          description: Estimated download size
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tools/colorize:
    post:
      tags: [Tools]
      summary: Colorize
      description: |
        AI-powered photo colorization. Converts black-and-white or grayscale
        photographs to full color using DDColor with OpenCV DNN fallback. Requires
        the colorize feature bundle to be installed.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Grayscale or B&W image file
                settings:
                  type: string
                  description: |
                    JSON string with options:
                    - `intensity` (number 0-1, default 1.0) - Colorization intensity
                    - `model` (string, default "auto") - One of: auto, ddcolor, opencv
                clientJobId:
                  type: string
                  description: Client-provided job ID for SSE progress tracking
      responses:
        "200":
          description: Colorized image
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ToolResponse"
                  - type: object
                    properties:
                      previewUrl:
                        type: string
                        description: WebP preview URL for non-previewable formats
                      width:
                        type: integer
                      height:
                        type: integer
                      method:
                        type: string
                        description: Colorization model used (ddcolor or opencv)
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "422":
          description: Processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Feature not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureNotInstalledError"
...
  /api/v1/tools/blur-faces:
    post:
      tags: [Tools]
      summary: Face blur
      description: Detect and blur faces in an image for privacy. Uses OpenCV. Runs locally.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, settings]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to process
                settings:
                  type: string
                  description: |
                    JSON string with options:
                    - `blurRadius` (number, default 30) — Blur strength radius
                    - `sensitivity` (number 0-1, default 0.5) — Face detection sensitivity
      responses:
        "200":
          description: Processed image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
...